refactor: move useRecentSessionHistory to a separate file
